requirements.txt
simplejson==3.7.3
Flask==0.10.1
Flask-restful==0.3.3
flask-mongoengine==0.7.1
mongoengine==0.10.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Cors==2.0.0
Flask-Mail==0.9.1
flask-marshmallow==0.6.0
marshmallow==2.0.0b3
marshmallow-mongoengine==0.6.4
PyJWT==1.1.0
passlib==1.6.2
Authomatic==0.0.13
requests==2.6.0
docopt==0.6.2
gunicorn==19.3.0 # Application server
xmltodict==0.9.2
pysolr==3.3.2
python-dateutil==2.4.2
python-daemon==2.0.5
mongopatcher==0.2.4
